# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "codestoon",
        "description": "codestoon api documantion",
        "version": "1.0.0"
    },
    "paths": {
        "/articles/categories": {
            "get": {
                "tags": [
                    "article category"
                ],
                "summary": "articles categories list",
                "description": "get list of the articles categories",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleCategoryController::index",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "collection",
                                            "example": "[{fa_title:'title',en_title:'title',description:'articlecategory description'},{fa_title:'title',en_title:'title',description:'articlecategory description'}]"
                                        },
                                        "message": {
                                            "type": "strig",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "article category"
                ],
                "summary": "store new article category",
                "description": "store new article category in databas",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleCategoryController::store",
                "requestBody": {
                    "description": "required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fa_title",
                                    "en_title",
                                    "description"
                                ],
                                "properties": {
                                    "fa_title": {
                                        "type": "string",
                                        "format": "article title",
                                        "example": "the article title in persian"
                                    },
                                    "en_title": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "the article title in english"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "article category description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/categories/{articleId}": {
            "put": {
                "tags": [
                    "article category"
                ],
                "summary": "update  article category",
                "description": "update  article category from databas",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleCategoryController::update",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "the articleId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fa_title",
                                    "en_title",
                                    "description"
                                ],
                                "properties": {
                                    "fa_title": {
                                        "type": "string",
                                        "format": "article title",
                                        "example": "the article title in persian"
                                    },
                                    "en_title": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "the article title in english"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "article category description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "article category"
                ],
                "summary": "delete  article category",
                "description": "delete  article category from databas",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleCategoryController::destroy",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "the articleId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/{article}": {
            "put": {
                "tags": [
                    "articles"
                ],
                "summary": "update article",
                "description": "update  article in the system",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleController::update",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "the article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response of the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "failed",
                                            "example": "failed"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "articles"
                ],
                "summary": "delete article",
                "description": "delete  article from the system",
                "operationId": "App\\Http\\Controllers\\api\\v1\\admin\\article\\ArticleController::destroy",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "the article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response of the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "failed",
                                            "example": "failed"
                                        },
                                        "data": {
                                            "type": "boolean",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}